feat(web): show index runtime inline
