+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added a live repository indexing job runtime popover to the syncing badge. [#1623](https://github.com/sourcebot-dev/sourcebot/pull/1623)
+- Added a live repository indexing job runtime to the syncing badge. [#1623](https://github.com/sourcebot-dev/sourcebot/pull/1623)
 
 ## [5.1.10] - 2026-08-27
