+++ packages/web/src/app/(app)/repos/components/reposTable.test.tsx
@@ -42,7 +42,7 @@ const repos: Repo[] = [
             id: "job-1",
             data: { repoId: 1 },
             status: "IN_PROGRESS",
-            startedAt: null,
+            startedAt: Date.now() - 30_000,
             errorMessage: null,
             result: null,
         },
@@ -290,7 +290,7 @@ describe("ReposTable", () => {
 
         await waitFor(() => {
             expect(reposActions.indexRepo).toHaveBeenCalledWith(2);
-            expect(screen.getByText("Syncing")).toBeTruthy();
+            expect(screen.getByText("Pending")).toBeTruthy();
             expect(fetch).toHaveBeenCalledOnce();
         });
 
@@ -313,7 +313,7 @@ describe("ReposTable", () => {
                 id: "active-reindex-job",
                 data: { repoId: repos[1].id },
                 status: "IN_PROGRESS",
-                startedAt: null,
+                startedAt: Date.now() - 30_000,
                 errorMessage: null,
                 result: null,
             },
@@ -546,7 +546,7 @@ describe("ReposTable", () => {
 
         await waitFor(() => {
             expect(reposActions.indexRepo).toHaveBeenCalledWith(1);
-            expect(screen.getByText("Syncing")).toBeTruthy();
+            expect(screen.getByText("Pending")).toBeTruthy();
             expect(fetch).toHaveBeenCalledOnce();
         });
         expect(screen.queryByText("Repository sync failed")).toBeNull();
@@ -607,7 +607,7 @@ describe("ReposTable", () => {
         expect(navigation.refresh).not.toHaveBeenCalled();
     });
 
-    test("polls a syncing repository whose latest job is missing", async () => {
+    test("polls a pending repository whose latest job is missing", async () => {
         const response: RepoIndexingStatusesResponse = {
             repositories: [{
                 repoId: 1,
@@ -627,10 +627,10 @@ describe("ReposTable", () => {
 
         renderTable([{ ...repos[0], latestJob: null }]);
 
-        expect(screen.getByText("Syncing")).toBeTruthy();
+        expect(screen.getByText("Pending")).toBeTruthy();
         await waitFor(() => expect(fetch).toHaveBeenCalledOnce());
         await waitFor(() => expect(screen.getByText("Failed")).toBeTruthy());
-        expect(screen.queryByText("Syncing")).toBeNull();
+        expect(screen.queryByText("Pending")).toBeNull();
         expect(navigation.refresh).not.toHaveBeenCalled();
     });
 });
